--- shared/src/persistence/createPolicyUrl.js
@@ -1,9 +1,0 @@
-/**
- * createUploadPolicy
- * @param applicationContext
- */
-exports.createUploadPolicy = ({ applicationContext }) => {
-  return  applicationContext.getPersistenceGateway().createUploadPolicy({
-    applicationContext,
-  });
-};
